Move NonInteractive upgrade to its own file
